requirements.txt
Flask==3.0.1
prettytable==3.9.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
waitress==2.1.2